+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 * migrated Android Plugin to new Flutter 1.12 API
 
+## Upcoming release
+* fix fetching device dnd for fresh installations
+
 ## 2.2.4
 * check partially fetched notifications when fetching inbox items