+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming release
 * fix fetching device dnd for fresh installations
 * fix custom params & headers on cloud API requests
+* fix beacons parsing when the result is empty
 
 ## 2.2.4
 * check partially fetched notifications when fetching inbox items