+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * fix user mapping
 * enhance NotificareUser model
 * fix handling new intents
+* fix scannable session
 
 ## 2.2.4
 * check partially fetched notifications when fetching inbox items